finish moving specs creation to class
